move utility functions to separate utils module
